services/project_discovery_service/requirements.txt
kafka-python>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0
PyGithub>=2.1.1
pydantic>=2.0.0
loguru>=0.7.0
requests>=2.31.0